requirements.txt
argparse ~= 1.4.0
flask ~= 3.0.2
lightgbm ~= 4.3.0
numpy ~= 1.26.4
pandas ~= 2.2.1
pyyaml ~= 6.0.1
scikit-learn ~= 1.4.1
waitress ~= 3.0.0
